Add support for multiple typing imports
