--- tests/example_files/expected_variable_definitions.py
@@ -1,41 +1,0 @@
-from collections import ChainMap
-from collections import OrderedDict
-from collections.abc import Mapping
-from collections import Counter
-from collections import defaultdict
-from collections.abc import Iterable
-from collections import deque
-import typing
-from typing import (
-)
-
-x: int
-x: float
-x: bool
-x: str
-x: bytes
-x: list
-x: list
-x: set
-x: set
-x: dict
-x: dict
-x: tuple
-x: tuple
-x: Iterable
-x: Iterable
-x: Mapping
-x: Mapping
-x: frozenset
-x: frozenset
-x: type
-x: type
-x: deque
-x: deque
-x: defaultdict
-x: defaultdict
-x: OrderedDict
-x: OrderedDict
-x: Counter
-x: Counter
-x: ChainMap
