+++ tests/integration/test_format_function_defintion.py
@@ -2,9 +2,6 @@ import os
 
 from tests import file_path
 
-example_file = file_path / 'example_files/expected_function_definition.py'
-project_dir = file_path.parent
-
 function_definition = '''
 import re
 import typing
@@ -59,20 +56,83 @@ def very_complex_function(
     return re.compile(f'{a},{b},{c},{d},{e},{f},{g},{h},{i},{j},{k},{l},{m}')
 '''
 
+expectd = '''
+from collections.abc import MutableSet
+from collections.abc import MutableMapping
+from collections.abc import Sequence
+from collections import OrderedDict
+from collections.abc import MappingView
+from collections.abc import Coroutine
+from collections.abc import Iterator
+from collections.abc import Iterable
+from contextlib import AbstractContextManager
+from contextlib import AbstractAsyncContextManager
+from collections.abc import AsyncGenerator
+from re import Pattern
+from collections.abc import ValuesView
+from collections.abc import ItemsView
+from collections.abc import Mapping
+from collections.abc import Collection
+from collections import ChainMap
+from re import Match
+from collections.abc import Callable
+from collections.abc import ByteString
+from collections import deque
+from collections.abc import KeysView
+from collections.abc import Reversible
+from collections.abc import AsyncIterable
+from collections.abc import MutableSequence
+from collections.abc import Set
+from collections.abc import Awaitable
+from collections import defaultdict
+from collections import Counter
+from collections.abc import Container
+from collections.abc import Generator
+from collections.abc import AsyncIterator
+
+import re
+import typing
+from typing import (
+    Union,
+)
+
+
+def very_complex_function(
+    a: tuple[list[dict[set[frozenset[type, deque]]]]],
+    b: defaultdict[OrderedDict[Counter, ChainMap]],
+    c: Awaitable,
+    d: Coroutine,
+    e: AsyncIterable[AsyncIterator[AsyncGenerator[Iterable, Iterator]]],
+    f: Generator[Reversible],
+    g: typing.Union[Callable, Set],
+    h: typing.Union[MutableSet[Mapping, MutableMapping], Sequence],
+    i: MutableSequence[Container[Collection]],
+    j: ByteString,
+    k: MappingView[KeysView[ItemsView[ValuesView]]],
+    l: AbstractContextManager,
+    m: typing.Optional[AbstractAsyncContextManager],
+) -> Union[Pattern, Match]:
+    return re.compile(f'{a},{b},{c},{d},{e},{f},{g},{h},{i},{j},{k},{l},{m}')
+'''
+
+result_path = file_path / 'result.py'
+expected_path = file_path / 'expected.py'
+
 
 def test_format_function_definition():
-    path = file_path / 'temp/test_format_function.py'
-    with open(path, 'w+') as file:
+    with open(result_path, 'w+') as file:
         file.write(function_definition)
+    with open(expected_path, 'w+') as file:
+        file.write(expectd)
 
     # Execute the pre-commit hook as a CLI
-    os.system(f'poetry run upgrade-type-hints-script {path}')
+    os.system(f'poetry run upgrade-type-hints-script {result_path}')
 
     # Load the changed file
-    with open(path, 'rb') as f:
+    with open(result_path, 'rb') as f:
         # we need to sort the lines, since the imports get added randomly atm
         result = sorted(f.readlines())
-    with open(example_file, 'rb') as f:
+    with open(expected_path, 'rb') as f:
         expected = sorted(f.readlines())
 
     result = [i for i in result if i.replace(b'\n', b'')]
@@ -80,22 +140,25 @@ def test_format_function_definition():
 
     for a, b in zip(result, expected):
         assert a == b
-    os.remove(path)
+
+    os.remove(result_path)
+    os.remove(expected_path)
 
 
 def test_format_function_definition_with_futures():
-    path = file_path / 'temp/test_format_function.py'
-    with open(path, 'w+') as file:
+    with open(result_path, 'w+') as file:
         file.write(function_definition)
+    with open(expected_path, 'w+') as file:
+        file.write(expectd)
 
     # Execute the pre-commit hook as a CLI
-    os.system(f'poetry run upgrade-type-hints-script {path} --futures=true')
+    os.system(f'poetry run upgrade-type-hints-script {result_path} --futures=true')
 
     # Load the changed file
-    with open(path, 'rb') as f:
+    with open(result_path, 'rb') as f:
         # we need to sort the lines, since the imports get added randomly atm
         result = sorted(f.readlines())
-    with open(example_file, 'rb') as f:
+    with open(expected_path, 'rb') as f:
         expected = sorted(f.readlines() + [b'from __future__ import annotations\n'])
 
     result = [i for i in result if i.replace(b'\n', b'')]
@@ -103,4 +166,6 @@ def test_format_function_definition_with_futures():
 
     for a, b in zip(result, expected):
         assert a == b
-    os.remove(path)
+
+    os.remove(result_path)
+    os.remove(expected_path)
