+++ tests/integration/test_format_variable_definitions.py
@@ -2,9 +2,6 @@ import os
 
 from tests import file_path
 
-example_file = file_path / 'example_files/expected_variable_definitions.py'
-project_dir = file_path.parent
-
 function_definition = '''
 import typing
 from typing import ChainMap, Counter
@@ -54,20 +51,69 @@ x: typing.Counter
 x: ChainMap
 '''
 
+expectd = '''
+from collections import ChainMap
+from collections import OrderedDict
+from collections.abc import Mapping
+from collections import Counter
+from collections import defaultdict
+from collections.abc import Iterable
+from collections import deque
+import typing
+from typing import (
+)
+
+x: int
+x: float
+x: bool
+x: str
+x: bytes
+x: list
+x: list
+x: set
+x: set
+x: dict
+x: dict
+x: tuple
+x: tuple
+x: Iterable
+x: Iterable
+x: Mapping
+x: Mapping
+x: frozenset
+x: frozenset
+x: type
+x: type
+x: deque
+x: deque
+x: defaultdict
+x: defaultdict
+x: OrderedDict
+x: OrderedDict
+x: Counter
+x: Counter
+x: ChainMap
+'''
+
+result_path = file_path / 'result.py'
+expected_path = file_path / 'expected.py'
+
 
 def test_format_variables():
-    path = file_path / 'temp/test_format_variables.py'
-    with open(path, 'w+') as file:
+    with open(result_path, 'w+') as file:
         file.write(function_definition)
 
+    with open(expected_path, 'w+') as file:
+        file.write(expectd)
+
     # Execute the pre-commit hook as a CLI
-    os.system(f'poetry run upgrade-type-hints-script {path}')
+    os.system(f'poetry run upgrade-type-hints-script {result_path}')
 
     # Load the changed file
-    with open(path, 'rb') as f:
+    with open(result_path, 'rb') as f:
         # we need to sort the lines, since the imports get added randomly atm
         result = sorted(f.readlines())
-    with open(example_file, 'rb') as f:
+    with open(expected_path, 'rb') as f:
         expected = sorted(f.readlines())
 
     result = [i for i in result if i.replace(b'\n', b'')]
@@ -76,22 +122,25 @@ def test_format_variables():
     for a, b in zip(result, expected):
         assert a == b
 
-    os.remove(path)
+    os.remove(result_path)
+    os.remove(expected_path)
 
 
 def test_format_variables_with_futures():
-    path = file_path / 'temp/test_format_variables.py'
-    with open(path, 'w+') as file:
+    with open(result_path, 'w+') as file:
         file.write(function_definition)
 
+    with open(expected_path, 'w+') as file:
+        file.write(expectd)
+
     # Execute the pre-commit hook as a CLI
-    os.system(f'poetry run upgrade-type-hints-script {path} --futures=true')
+    os.system(f'poetry run upgrade-type-hints-script {result_path} --futures')
 
     # Load the changed file
-    with open(path, 'rb') as f:
+    with open(result_path, 'rb') as f:
         # we need to sort the lines, since the imports get added randomly atm
         result = sorted(f.readlines())
-    with open(example_file, 'rb') as f:
+    with open(expected_path, 'rb') as f:
         expected = sorted(f.readlines() + [b'from __future__ import annotations\n'])
 
     result = [i for i in result if i.replace(b'\n', b'')]
@@ -100,4 +149,5 @@ def test_format_variables_with_futures():
     for a, b in zip(result, expected):
         assert a == b
 
-    os.remove(path)
+    os.remove(result_path)
+    os.remove(expected_path)
